+++ packages/core/src/index.ts
@@ -1,4 +1,4 @@
-export const VERSION = "0.6.34";
+export const VERSION = "0.6.35";
 
 export type {
 	CreateComponentArgs,